fixed render float/int bugs and successful first sample program for entities

+++ include/core/util/Entity.h
@@ -21,7 +21,7 @@ class Entity {
 
   int getID() const;
 
-  virtual void render(Vector2D positionOffset, int ratio) const = 0;
+  virtual void render(Vector2D positionOffset, float ratio) const = 0;
   
 };
